refactor: move app and animation files to own directories
