+++ src/apps/bpmCalculator.ts
