+++ src/content/apps/bpm-calculator.mdx
@@ -14,7 +14,7 @@ keywords:
   - "note duration"
 ---
 
-import BpmCalculator from '../../components/apps/BpmCalculator.svelte';
+import BpmCalculator from '../../apps/BpmCalculator.svelte';
 
 <BpmCalculator client:load />
