+++ src/content/media/ocean-waves.mdx
@@ -22,7 +22,7 @@ keywords:
   - procedural visualization
 ---
 
-import OceanWaves from '../../components/apps/OceanWaves.svelte';
+import OceanWaves from '../../animations/OceanWaves.svelte';
 
 <OceanWaves client:load height={600} audioUrl={props.audioUrl} />
